First implementation of the player controller. Partially functional controller scheme and it's inspector.
This commint, introduces the firts concep of the controller scheme. The ControllerScheme is the set of instructions a PlayerController uses to execute commands, which  in turn, can be attached and dettached via inspector to facilitate customization and testing of binds, as also, it can be done through code, allowing customizaiton in run time. Still conceptual though.